refactor each entity type into its own file, move higher-level entities into their own files too, update imports everywhere
