+++ postgres_declare/entities/base_entity.py
