+++ tests/it/test_database.py
@@ -3,8 +3,9 @@ from hypothesis import HealthCheck, given, settings
 from hypothesis import strategies as st
 from sqlalchemy import Engine
 
-from postgres_declare.base_entity import Entity
-from postgres_declare.cluster_entities import Database, Role
+from postgres_declare.entities.base_entity import Entity
+from postgres_declare.entities.database import Database
+from postgres_declare.entities.role import Role
 
 
 def test_does_not_exist(simple_db: Database) -> None:
