+++ tests/it/test_schema.py
@@ -1,9 +1,10 @@
 import pytest
 from sqlalchemy import Engine
 
-from postgres_declare.base_entity import Entity
-from postgres_declare.cluster_entities import Database, Role
-from postgres_declare.database_entities import Schema
+from postgres_declare.entities.base_entity import Entity
+from postgres_declare.entities.database import Database
+from postgres_declare.entities.role import Role
+from postgres_declare.entities.schema import Schema
 
 
 def test_does_not_exist(simple_schema: Schema, simple_db: Database) -> None:
